requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
tkcalendar>=1.6.1
pyinstaller>=6.0.0
pillow>=10.0.0
python-docx>=1.1.0
reportlab>=4.0.0
